chore: refactor root views
