+++ oogway/lib/src/ui/root/root_view.dart
@@ -2,6 +2,9 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:oogway/src/common/constants/ui.dart';
 import 'package:oogway/src/ui/root/controllers/index_controller.dart';
+import 'package:oogway/src/ui/root/ui/home/home_view.dart';
+import 'package:oogway/src/ui/root/ui/search/search_view.dart';
+import 'package:oogway/src/ui/root/ui/settings/settings_view.dart';
 
 class RootView extends ConsumerWidget {
   const RootView({Key? key}) : super(key: key);
@@ -37,9 +40,9 @@ class RootView extends ConsumerWidget {
       body: IndexedStack(
         index: bottomNav,
         children: const [
-          Center(child: Text("Home")),
-          Center(child: Text("Search")),
-          Center(child: Text("Settings")),
+          HomeView(),
+          SearchView(),
+          SettingsView(),
         ],
       ),
       bottomNavigationBar: BottomNavigationBar(
