+++ oogway/lib/src/ui/root/ui/search/search_view.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/material.dart';
+
+class SearchView extends StatelessWidget {
+  const SearchView({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return const Center(
+      child: Text('Search'),
+    );
+  }
+}
